KTOR-2214 Fix autoreload in embeddedServer
